--- a/.gitignore
+++ b/.gitignore
@@ -4,6 +4,7 @@ build
 dist
 extlib_*/
 scratch_*
+perf/results/**
 
 # Created by https://www.toptal.com/developers/gitignore/api/python,macos,emacs,cmake,virtualenv
 # Edit at https://www.toptal.com/developers/gitignore?templates=python,macos,emacs,cmake,virtualenv