--- a/pyproject.toml
+++ b/pyproject.toml
@@ -66,6 +66,11 @@ dev = [
     "boto3>=1.35.67,<2.0.0",
     "autoflake>=2.3.1,<3.0.0"
 ]
+perf-test = [
+    "pdfplumber>=0.11.7",
+    "pymupdf>=1.26.4",
+    "pypdfium2>=4.30.0",
+]
 [tool.uv]
 package = true
 default-groups = "all"