--- a/pyproject.toml
+++ b/pyproject.toml
@@ -71,6 +71,13 @@ perf-test = [
     "pymupdf>=1.26.4",
     "pypdfium2>=4.30.0",
 ]
+
+[project.optional-dependencies]
+perf-tools = [
+    "pdfplumber>=0.11.7",
+    "pymupdf>=1.26.4",
+    "pypdfium2>=4.30.0",
+]
 [tool.uv]
 package = true
 default-groups = "all"