We are adding perf-tools to compare native docling parse with other parsing libraries.

```
taa@Munlochy docling-parse % uv run python perf/run_perf.py /Users/taa/Documents/projects/_data/cas_gt20 --parser pypdfium2
Parsing PDFs with pypdfium2: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20/20 [00:09<00:00,  2.17it/s]

Summary for parser=pypdfium2
 - files:        20
 - pages total:  2270
 - pages ok:     2270
 - pages failed: 0
 - total sec:    9.071300
 - avg sec/page: 0.003996
 - p50: 0.002912  p90: 0.004833  p95: 0.007059  p99: 0.042817
 - min: 0.000027  max: 0.105380

Per-document statistics (sec/page):
document                                                                                                                     pages     total      mean    median       min       max       p90       p95       p99
-------------------------------------------------------------------------------------------------------------------------  -------  --------  --------  --------  --------  --------  --------  --------  --------
/Users/taa/Documents/projects/_data/cas_gt20/cd-bitmap-space-configuration-2.pdf                                                 3  0.023439  0.007813  0.008968  0.005376  0.009096  0.00907   0.009083  0.009093
/Users/taa/Documents/projects/_data/cas_gt20/concurrent-compatibility-and-code-cross-reference-ibm-storage-virtualize.pdf       41  0.369759  0.009019  0.006148  0.001466  0.059231  0.010384  0.013811  0.056612
/Users/taa/Documents/projects/_data/cas_gt20/configuring-call-home.pdf                                                           1  0.005256  0.005256  0.005256  0.005256  0.005256  0.005256  0.005256  0.005256
/Users/taa/Documents/projects/_data/cas_gt20/pip-environmental-requirements-3.pdf                                                5  0.035835  0.007167  0.007101  0.004275  0.010678  0.009393  0.010035  0.010549
/Users/taa/Documents/projects/_data/cas_gt20/pools-creating-data-reduction.pdf                                                   3  0.021791  0.007264  0.009914  0.001322  0.010555  0.010427  0.010491  0.010543
/Users/taa/Documents/projects/_data/cas_gt20/problem-procedure-understanding-system-status-from-leds.pdf                         6  0.030994  0.005166  0.005928  0.001246  0.006533  0.006292  0.006412  0.006509
/Users/taa/Documents/projects/_data/cas_gt20/redp5520.pdf                                                                       16  0.054211  0.003388  0.002783  3e-05     0.013255  0.005441  0.007547  0.012113
/Users/taa/Documents/projects/_data/cas_gt20/redp5534.pdf                                                                      118  0.395341  0.00335   0.00222   2.8e-05   0.10538   0.004422  0.004927  0.007827
/Users/taa/Documents/projects/_data/cas_gt20/redp5613.pdf                                                                      152  0.514253  0.003383  0.003137  2.9e-05   0.037295  0.004338  0.007336  0.01078
/Users/taa/Documents/projects/_data/cas_gt20/redp5617.pdf                                                                       72  0.205231  0.00285   0.002802  3.2e-05   0.008208  0.004025  0.004257  0.006277
/Users/taa/Documents/projects/_data/cas_gt20/redp5638.pdf                                                                       54  0.128385  0.002378  0.002439  2.8e-05   0.008159  0.003402  0.003866  0.006801
/Users/taa/Documents/projects/_data/cas_gt20/redp5669.pdf                                                                       74  0.208506  0.002818  0.002815  2.9e-05   0.008052  0.004042  0.004442  0.006295
/Users/taa/Documents/projects/_data/cas_gt20/redp5670.pdf                                                                       48  0.145144  0.003024  0.003097  2.7e-05   0.008027  0.003985  0.004731  0.006918
/Users/taa/Documents/projects/_data/cas_gt20/sg248468.pdf                                                                      452  1.06892   0.002365  0.002271  2.8e-05   0.007948  0.003902  0.004371  0.005224
/Users/taa/Documents/projects/_data/cas_gt20/sg248497.pdf                                                                      364  1.0194    0.002801  0.002781  2.7e-05   0.011013  0.004282  0.004863  0.005895
/Users/taa/Documents/projects/_data/cas_gt20/sg248521.pdf                                                                      784  2.38926   0.003048  0.003148  2.7e-05   0.008834  0.004429  0.004843  0.005684
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_6172647.pdf                                                      4  0.082807  0.020702  0.019703  0.001479  0.041922  0.036375  0.039149  0.041367
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_6380846.pdf                                                      3  0.071179  0.023726  0.013885  0.007232  0.050061  0.042826  0.046444  0.049338
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_690527.pdf                                                       6  0.094528  0.015755  0.015209  0.003775  0.028282  0.025911  0.027096  0.028045
/Users/taa/Documents/projects/_data/cas_gt20/system-storage-ts3500-tape-library-model-d23.pdf                                   64  2.20705   0.034485  0.02846   0.002481  0.079582  0.062871  0.070184  0.077868

Wrote: perf/results/perf_pypdfium2_20250915-163339.csv
Total wall time: 9.215333 sec
taa@Munlochy docling-parse % uv run python perf/run_perf.py /Users/taa/Documents/projects/_data/cas_gt20 --parser docling
Parsing PDFs with docling: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 20/20 [00:56<00:00,  2.83s/it]

Summary for parser=docling
 - files:        20
 - pages total:  2270
 - pages ok:     2270
 - pages failed: 0
 - total sec:    55.937700
 - avg sec/page: 0.024642
 - p50: 0.018701  p90: 0.047992  p95: 0.057145  p99: 0.138790
 - min: 0.000082  max: 0.695331

Per-document statistics (sec/page):
document                                                                                                                     pages      total      mean    median       min       max       p90       p95       p99
-------------------------------------------------------------------------------------------------------------------------  -------  ---------  --------  --------  --------  --------  --------  --------  --------
/Users/taa/Documents/projects/_data/cas_gt20/cd-bitmap-space-configuration-2.pdf                                                 3   0.066386  0.022129  0.017372  0.016411  0.032603  0.029557  0.03108   0.032299
/Users/taa/Documents/projects/_data/cas_gt20/concurrent-compatibility-and-code-cross-reference-ibm-storage-virtualize.pdf       41   0.326398  0.007961  0.006582  0.002146  0.030672  0.011083  0.01489   0.028812
/Users/taa/Documents/projects/_data/cas_gt20/configuring-call-home.pdf                                                           1   0.019819  0.019819  0.019819  0.019819  0.019819  0.019819  0.019819  0.019819
/Users/taa/Documents/projects/_data/cas_gt20/pip-environmental-requirements-3.pdf                                                5   0.103324  0.020665  0.023853  0.007514  0.032654  0.029866  0.03126   0.032375
/Users/taa/Documents/projects/_data/cas_gt20/pools-creating-data-reduction.pdf                                                   3   0.091617  0.030539  0.029676  0.00262   0.059322  0.053393  0.056357  0.058729
/Users/taa/Documents/projects/_data/cas_gt20/problem-procedure-understanding-system-status-from-leds.pdf                         6   0.067737  0.011289  0.011884  0.002055  0.016188  0.015311  0.01575   0.016101
/Users/taa/Documents/projects/_data/cas_gt20/redp5520.pdf                                                                       16   0.314556  0.01966   0.016715  8.2e-05   0.087908  0.030249  0.044779  0.079282
/Users/taa/Documents/projects/_data/cas_gt20/redp5534.pdf                                                                      118   2.34795   0.019898  0.012499  0.000118  0.138592  0.047552  0.059891  0.11705
/Users/taa/Documents/projects/_data/cas_gt20/redp5613.pdf                                                                      152   4.62904   0.030454  0.017637  0.000134  0.695331  0.052979  0.105751  0.148997
/Users/taa/Documents/projects/_data/cas_gt20/redp5617.pdf                                                                       72   1.65364   0.022967  0.022107  0.000114  0.087608  0.035781  0.049695  0.064743
/Users/taa/Documents/projects/_data/cas_gt20/redp5638.pdf                                                                       54   0.763623  0.014141  0.012112  8.3e-05   0.08701   0.022073  0.026895  0.083627
/Users/taa/Documents/projects/_data/cas_gt20/redp5669.pdf                                                                       74   1.7683    0.023896  0.022534  0.000102  0.087034  0.041909  0.046114  0.064978
/Users/taa/Documents/projects/_data/cas_gt20/redp5670.pdf                                                                       48   1.31494   0.027395  0.025007  0.000106  0.086364  0.046853  0.051582  0.071403
/Users/taa/Documents/projects/_data/cas_gt20/sg248468.pdf                                                                      452   9.09468   0.020121  0.012579  0.000187  0.169903  0.048343  0.060521  0.086648
/Users/taa/Documents/projects/_data/cas_gt20/sg248497.pdf                                                                      364   9.25975   0.025439  0.020284  0.000221  0.168398  0.048626  0.055955  0.137817
/Users/taa/Documents/projects/_data/cas_gt20/sg248521.pdf                                                                      784  22.6357    0.028872  0.023877  0.000341  0.178095  0.050237  0.057951  0.166556
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_6172647.pdf                                                      4   0.061035  0.015259  0.016006  0.00249   0.026534  0.024042  0.025288  0.026284
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_6380846.pdf                                                      3   0.050379  0.016793  0.013387  0.009066  0.027926  0.025018  0.026472  0.027635
/Users/taa/Documents/projects/_data/cas_gt20/support_pages_node_690527.pdf                                                       6   0.080042  0.01334   0.014352  0.004538  0.021341  0.019659  0.0205    0.021173
/Users/taa/Documents/projects/_data/cas_gt20/system-storage-ts3500-tape-library-model-d23.pdf                                   64   1.2888    0.020137  0.016832  0.00263   0.042742  0.033855  0.038352  0.041713

Wrote: perf/results/perf_docling_20250915-163355.csv
Total wall time: 56.524318 sec
````